Added pytest_plugins in conftest
